+++ source/extensions/omni.isaac.core/docs/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 Added
 ------
 - An argument to clear scene registery only
+- rotation and cross product util functions
 
 Fixed
 ------